requirements.txt
aiohttp==3.11.11
attr==0.3.2
ConfigParser==7.1.0
cryptography==42.0.5
Cython==3.0.11
dl==0.1.0
docutils==0.21.2
h2==4.1.0
HTMLParser==0.0.2
importlib_metadata==8.5.0
ipython==8.12.3
ipywidgets==8.1.5
Jinja2==3.1.5
js==1.0
keyring==25.5.0
lockfile==0.12.2
lxml==5.3.0
numpy==2.2.1
ordereddict==1.1
openai
pandas==2.2.3
Pillow==11.0.0
plotly==5.24.1
protobuf==5.29.2
pyOpenSSL==24.3.0
pytest==8.3.4
python-dotenv
PyYAML==6.0.2  # if error, try pyyaml
railroad==0.5.0
simplejson==3.19.3
Sphinx==8.1.3
streamlit==1.41.1
thread==2.0.5
toml==0.10.2
tornado==6.4.2
zipp==3.20.2
zstandard==0.22.0
